Photo Capabilities
Added the ability to edit the recipe photo by taking a photo form the
phone. Added the ability to add a photo when creating a recipe. Added
image views that display the item photos.

Loads data from Parse but doesn’t save anything since I don’t have
access to the Parse account and don’t want to mess with it.